+++ packages/component/src/ScrollToBottom/Panel.js
@@ -1,21 +1,22 @@
-import { css } from 'glamor';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import React, { useContext } from 'react';
 
 import InternalContext from './InternalContext';
+import useStyleToClassName from '../hooks/internal/useStyleToClassName';
 
-const ROOT_CSS = css({
+const ROOT_STYLE = {
   height: '100%',
   overflowY: 'auto',
   width: '100%'
-});
+};
 
 const Panel = ({ children, className }) => {
   const { setTarget } = useContext(InternalContext);
+  const rootCSS = useStyleToClassName()(ROOT_STYLE);
 
   return (
-    <div className={classNames(ROOT_CSS + '', (className || '') + '')} ref={setTarget}>
+    <div className={classNames(rootCSS + '', (className || '') + '')} ref={setTarget}>
       {children}
     </div>
   );