+++ packages/component/src/hooks/internal/useInternalContext.js
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import InternalContext from '../../ScrollToBottom/InternalContext';
+
+export default function useInternalContext() {
+  return useContext(InternalContext);
+}