+++ packages/component/src/hooks/internal/useStateContext.js
@@ -1,0 +1,7 @@
+import { useContext } from 'react';
+
+import StateContext from '../../ScrollToBottom/StateContext';
+
+export default function useStateContext() {
+  return useContext(StateContext);
+}