+++ packages/component/src/hooks/useAnimatingToEnd.js
@@ -1,9 +1,7 @@
-import { useContext } from 'react';
-
-import StateContext from '../ScrollToBottom/StateContext';
+import useStateContext from './internal/useStateContext';
 
 export default function useAnimatingToEnd() {
-  const context = useContext(StateContext);
+  const { animatingToEnd } = useStateContext();
 
-  return [context.animatingToEnd];
+  return [animatingToEnd];
 }