+++ packages/component/src/hooks/useAtEnd.js
@@ -1,9 +1,7 @@
-import { useContext } from 'react';
-
-import StateContext from '../ScrollToBottom/StateContext';
+import useStateContext from './internal/useStateContext';
 
 export default function useAtEnd() {
-  const context = useContext(StateContext);
+  const { atEnd } = useStateContext();
 
-  return [context.atEnd];
+  return [atEnd];
 }