+++ packages/playground/src/CommandBar.js
@@ -1,4 +1,4 @@
-import { css } from 'glamor';
+import { css } from 'emotion';
 import classNames from 'classnames';
 import React, { useCallback, useRef } from 'react';